Reduced boilerplate code for certain operations

+++ app/src/main/java/com/studio/confidoworks/clicker/variables.java
@@ -24,13 +24,14 @@ package com.studio.confidoworks.clicker;
 class variables
 {
     static boolean run;
-    static int clicks;
-    static int cps;
-    static int clicksPC;
+    static long clicks;
+    static long cps;
+    static long clicksPC;
+
     static int store1Num;
-    static int store1Price;
+    static long store1Price;
     static int store2Num;
-    static int store2Price;
+    static long store2Price;
     static int store3Num;
-    static int store3Price;
+    static long store3Price;
 }
